Move ItemCell into owned file
